refactor: extract startup shutdown lifecycle boundary
