+++ rustfs/src/lib.rs
@@ -78,6 +78,7 @@ pub mod startup_protocols;
 pub mod startup_runtime;
 pub mod startup_server;
 pub mod startup_services;
+pub mod startup_shutdown;
 pub mod startup_storage;
 pub mod storage;
 pub(crate) mod storage_compat;
